feat: rename view to quickView
