+++ src/components/pool/quickView.less
